data/generation/pyproject.toml
[tool.poetry]
name = "retail-rag-data-ingestion"
version = "0.1.0"
description = "Data generation and ingestion for the Retail RAG project."
authors = ["Siddharth Nandagopal"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.13.0"
psycopg2-binary = "^2.9"
faker = "^18.9"
python-dotenv = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
